+++ src/GeoMakie.jl
@@ -68,8 +68,6 @@ include("makie-axis.jl")
 include("mesh_image.jl")
 include("linesplitting.jl")
 
-include("polar.jl") # GeoPolarAxis: pole-centred azimuthal maps on a Makie PolarAxis
-
 include("sphere/unit_sphere_transforms.jl")
 include("sphere/icosphere.jl")
 include("sphere/globetransform.jl")
@@ -82,7 +80,7 @@ export Proj
 
 export FileIO
 
-export GeoAxis, GeoPolarAxis, add_cyclic_point, automatic
+export GeoAxis, add_cyclic_point, automatic
 export datalims, datalims!
 @deprecate datalims Makie.autolimits
 @deprecate datalims! Makie.reset_limits!