+++ test/runtests.jl
@@ -22,6 +22,6 @@ Makie.set_theme!(Theme(
     @testset "GeoAxisV2" include("geoaxis_v2.jl")
     @testset "MeshImage" include("meshimage.jl")
     @testset "GeoAxis" include("geoaxis.jl")
-    @testset "GeoPolarAxis" include("polar.jl")
+    @testset "PolarGeoAxis" include("polar_geoaxis.jl")
     @testset "GlobeAxis" include("globeaxis.jl")
 end